header com botões todos componetizados, esperando estados para condicionar
