+++ pokedex/src/Router/Router.js
@@ -1,7 +1,5 @@
 import { BrowserRouter, Switch, Route} from "react-router-dom"
-import HeaderHome from '../Components/HeaderHome'
-import HeaderDetails from '../Components/HeaderDetails'
-import HeaderPokedex from '../Components/HeaderPokedex'
+import Header from '../Components/Header'
 import Home from '../Pages/Home/Home'
 import Pokedex from '../Pages/Pokedex/Pokedex'
 import Details from '../Pages/Details/Details'
@@ -9,19 +7,17 @@ import Details from '../Pages/Details/Details'
 const Router = () => {
     return(
     <BrowserRouter>
+      <Header/>
         <Switch>
             <Route exact path="/">
-                <HeaderHome/>
                 <Home/>
             </Route>
 
             <Route exact path="/pokedex">
-                <HeaderPokedex/>
                 <Pokedex/>
             </Route>
 
             <Route exact path="/details">
-                <HeaderDetails/>
                 <Details/>
             </Route>
         </Switch>
